% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkItemResponse.R
\name{itemHierarchy}
\alias{itemHierarchy}
\title{Construct item names that capture the hierarchical of an ODK questionnare.}
\usage{
itemHierarchy(odk_form)
}
\arguments{
\item{odk_form}{A data frame containing an ODK form.}
}
\description{
\code{itemHierarchy} map the ODK questionnaire hierarchy (i.e. a replication
of the field names generated by ODK Aggregate and used in the CSV export).
}
\details{
Returns the the same data frame, but with an additional column
consisting of the group names each item belongs to.  The group names
are separated by ".".
}
\examples{
\dontrun{
## Example with 2016 WHO VA instrument version 1.5.1

form_f_name151 <- system.file("forms",
                              "WHOVA2016_v1_5_1_XLS_form_for_ODK_survey.csv",
                              package = "CrossVA")
form151 <- read.csv2(form_f_name151, stringsAsFactors = FALSE)
item_groups <- itemHierarchy(form151)$item_groups
}

}
