% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkItemResponse.R
\name{itemMissing}
\alias{itemMissing}
\title{Calculate item missingness.}
\usage{
itemMissing(odk_data, odk_form, id_col = "meta.instanceID")
}
\arguments{
\item{odk_data}{A data frame with ODK data}

\item{odk_form}{A data frame containing an ODK form.}

\item{id_col}{A string with column name in odk_data with ID (default is "meta.instanceID")}
}
\description{
\code{itemMissing}
}
\details{
List with two data frames Deaths and Items.
}
\examples{
\dontrun{
## Example with 2016 WHO VA instrument version 1.5.1

record_f_name151 <- system.file("sample",
                                "who151_odk_export.csv",
                                package = "CrossVA")
records151 <- read.csv(record_f_name151, stringsAsFactors = FALSE)


form_f_name151 <- system.file("forms",
                              "WHOVA2016_v1_5_1_XLS_form_for_ODK_survey.csv",
                              package = "CrossVA")
form151 <- read.csv2(form_f_name151, stringsAsFactors = FALSE)
results <- itemMissing(records151, form151)
}

}
