% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkItemResponse.R
\name{translate}
\alias{translate}
\title{Translate ODK relevant field}
\usage{
translate(relevant, death)
}
\arguments{
\item{relevant}{A string from ODK$relevant that needs to be translated.}

\item{death}{A dataframe (one death) of ODK data.}
}
\description{
\code{translate} converts conditions in ODK to R.
}
\details{
This is a worker function.
}
\examples{
\dontrun{
## Example with 2016 WHO VA instrument version 1.5.1
odkData <- read.csv("who151_odk_export.csv", stringsAsFactors = FALSE)
names(odkData)
odkForm <- read_xlsx("08_1_WHOVA2016_v1_5_1_XLS_form_for_ODK.xlsx", sheet = "survey")
}

}
