% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk2openVA_v151.R
\name{odk2openVA_v151}
\alias{odk2openVA_v151}
\title{Map VA records (version 1.5.1) to InterVA5 & InSilico (with option data.type = "WHO2016").}
\usage{
odk2openVA_v151(odk)
}
\arguments{
\item{odk}{A dataframe, obtained from reading an ODKBriefcase
export of records collected with the WHO questionnaire.}
}
\description{
\code{odk2openVA_v151} transform data collected with the 2016 WHO VA instrument (version 1.5.1)
  to serve different InterVA5 and InSilicoVA alogrithms for coding cause of death.
}
\examples{
\dontrun{
record_f_name <- system.file('sample', 'who151_va_output.csv', package = 'CrossVA')
records <- read.csv(record_f_name)
output_data_IV5 <- odk2openVA_v151(records)
}

}
