% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkItemResponse.R
\name{translate}
\alias{translate}
\title{Translate ODK relevant field}
\usage{
translate(relevant, death)
}
\arguments{
\item{relevant}{A string from ODK$relevant that needs to be translated.}

\item{death}{A dataframe (one death) of ODK data.}
}
\description{
\code{translate} converts conditions in ODK to R.
}
\details{
This is a worker function.
}
\examples{
\dontrun{
## Example with 2016 WHO VA instrument version 1.5.1
record_f_name151 <- system.file("sample",
                                "who151_odk_export.csv",
                                package = "CrossVA")
records151 <- read.csv(record_f_name151, stringsAsFactors = FALSE)

form_f_name151 <- system.file("forms",
                              "WHOVA2016_v1_5_1_XLS_form_for_ODK_survey.csv",
                              package = "CrossVA")
form151 <- read.csv2(form_f_name151, stringsAsFactors = FALSE)
death <- records151[1,]

relevant <- form151$relevant[15]
translate(relevant, death)
}

}
